+++ src/applications/SimulationResults.java
@@ -67,4 +67,28 @@ public class SimulationResults {
         jobCompletions[nextJob] = jobCompletionData;
         nextJob++;
     }
+
+    public static class JobCompletionData {
+        private final int completionTime;
+        private final int totalWaitTime;
+        private final int jobNumber;
+
+        public JobCompletionData(int jobNumber, int completionTime, int totalWaitTime) {
+            this.jobNumber = jobNumber;
+            this.completionTime = completionTime;
+            this.totalWaitTime = totalWaitTime;
+        }
+
+        public int getCompletionTime() {
+            return completionTime;
+        }
+
+        public int getTotalWaitTime() {
+            return totalWaitTime;
+        }
+
+        public int getJobNumber() {
+            return jobNumber;
+        }
+    }
 }
