+++ tests/properties/applications/SimulationProperties.java
@@ -19,7 +19,7 @@ public class SimulationProperties {
     {
         final SimulationResults results = MachineShopSimulator.runSimulation(specification);
         final int finishTime = results.getFinishTime();
-        final JobCompletionData[] jobCompletionData = results.getJobCompletionData();
+        final SimulationResults.JobCompletionData[] jobCompletionData = results.getJobCompletionData();
         final int lastJobCompletionTime = jobCompletionData[jobCompletionData.length-1].getCompletionTime();
         assertEquals(finishTime, lastJobCompletionTime);
     }
@@ -38,7 +38,7 @@ public class SimulationProperties {
         }
 
         int totalJobWaitTime = 0;
-        for (JobCompletionData jobCompletionData : results.getJobCompletionData()) {
+        for (SimulationResults.JobCompletionData jobCompletionData : results.getJobCompletionData()) {
             final int jobWaitTime = jobCompletionData.getTotalWaitTime();
             assertThat(jobWaitTime, greaterThanOrEqualTo(0));
             totalJobWaitTime += jobWaitTime;
@@ -54,7 +54,7 @@ public class SimulationProperties {
     {
         final SimulationResults results = MachineShopSimulator.runSimulation(specification);
 
-        JobCompletionData[] jobCompletionData = results.getJobCompletionData();
+        SimulationResults.JobCompletionData[] jobCompletionData = results.getJobCompletionData();
         for (int i=1; i<jobCompletionData.length-1; ++i) {
             assertThat(jobCompletionData[i].getCompletionTime(),
                     lessThanOrEqualTo(jobCompletionData[i+1].getCompletionTime()));
